Fixed audio seek bugs introduced in rev. 5117
svn path=/trunk/libavg/; revision=5124

+++ src/video/AudioDecoderThread.cpp
@@ -72,7 +72,7 @@ bool AudioDecoderThread::work()
     return true;
 }
 
-void AudioDecoderThread::seek(long long DestTime)
+void AudioDecoderThread::seek(double DestTime)
 {
     try {
         while (!m_MsgQ.empty()) {
